+++ src/main/java/com/aplbackend/pruebaTecnica/repositorio/ProductoRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aplbackend.pruebaTecnica.modelo.Producto;
 
-public interface ProductoRepositorio extends JpaRepository<Producto, Integer> {
+public interface ProductoRepositorio extends JpaRepository<Producto, Long> {
 
 }
