+++ src/main/java/com/aplbackend/pruebaTecnica/repositorio/UsuarioRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aplbackend.pruebaTecnica.modelo.Usuario;
 
-public interface UsuarioRepositorio extends JpaRepository<Usuario, Integer>{
+public interface UsuarioRepositorio extends JpaRepository<Usuario, Long>{
 
 }
