+++ src/main/java/com/aplbackend/pruebaTecnica/repositorio/VentaProductoRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aplbackend.pruebaTecnica.modelo.VentaProducto;
 
-public interface VentaProductoRepositorio extends JpaRepository<VentaProducto, Integer> {
+public interface VentaProductoRepositorio extends JpaRepository<VentaProducto, Long> {
 
 }
