+++ src/main/java/com/aplbackend/pruebaTecnica/repositorio/VentaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aplbackend.pruebaTecnica.modelo.Venta;
 
-public interface VentaRepositorio extends JpaRepository<Venta, Integer> {
+public interface VentaRepositorio extends JpaRepository<Venta, Long> {
 
 }
